Isolate validation functions and save token in local storage
