+++ src/components/login-mutation.ts
@@ -1,6 +1,16 @@
 import { gql } from '@apollo/client';
 
-export const LoginMutation = gql`
+type UserType = {
+  id: string;
+  name: string;
+};
+
+export type LoginType = {
+  token: string;
+  user: UserType;
+};
+
+export const loginMutation = gql`
   mutation LoginMutation($email: String!, $password: String!) {
     login(data: { email: $email, password: $password }) {
       token
